+++ connect/src/main/java/com/bau/connect/Whiteboard.java
@@ -7,15 +7,15 @@ import java.util.ArrayList;
  * Client uses when it gets an add shape message
  */
 public interface Whiteboard {
-    public void addPolygon(Integer id, Color c, ArrayList<Double> x, ArrayList<Double> y);
+    public void addPolygon(Integer id, Color c, ArrayList<Integer> x, ArrayList<Integer> y);
 
-    public void addText(Integer id, Color c, String s, Integer i, Double x, Double y, String text);
+    public void addText(Integer id, Color c, String s, Integer i, Integer x, Integer y, String text);
 
-    public void addLine(Integer id, Color c, Double x1, Double y1, Double x2, Double y2);
+    public void addLine(Integer id, Color c, Integer x1, Integer y1, Integer x2, Integer y2);
 
-	public void addRect(Integer id, Color c, Double x, Double y, Double w, Double h);
+	public void addRect(Integer id, Color c, Integer x, Integer y, Integer w, Integer h);
     
-	public void addOval(Integer id, Color c, Double x, Double y, Double w, Double h);
+	public void addOval(Integer id, Color c, Integer x, Integer y, Integer w, Integer h);
 	
 	public void removeShape(Integer id);
 }
